WiCS Restructure and Redesign
Re-designed WiCS for a more general and public applicability and
relevance, whereas it was previously designed and suited for internal
use within a given 'organizational' structure.

Migrated WiCS to Laravel 5

Updated the Architectural design, layout and structure of the
application to be more relevant and effective for the new solution.

--documentation to be formalized soon... :)